requirements.txt
requests==2.9.1
tweepy==3.5.0
python-bittrex
